+++ apps/prairielearn/src/tests/helperServer.ts
@@ -1,8 +1,8 @@
 import { setTimeout as sleep } from 'node:timers/promises';
 
-import { assert } from 'chai';
 import debugfn from 'debug';
 import * as tmp from 'tmp-promise';
+import { assert } from 'vitest';
 
 import { cache } from '@prairielearn/cache';
 import * as opentelemetry from '@prairielearn/opentelemetry';
@@ -28,8 +28,8 @@ const debug = debugfn('prairielearn:helperServer');
 const sql = sqldb.loadSqlEquiv(import.meta.url);
 
 config.startServer = false;
-// Pick a unique port based on the Mocha worker ID.
-config.serverPort = (3007 + Number.parseInt(process.env.MOCHA_WORKER_ID ?? '0', 10)).toString();
+// Pick a unique port based on the Vitest worker ID.
+config.serverPort = (3007 + Number.parseInt(process.env.VITEST_POOL_ID ?? '0', 10)).toString();
 
 export function before(courseDir: string | string[] = TEST_COURSE_PATH): () => Promise<void> {
   return async () => {
@@ -39,8 +39,7 @@ export function before(courseDir: string | string[] = TEST_COURSE_PATH): () => P
       await opentelemetry.init({ openTelemetryEnabled: false });
 
       debug('before(): initializing DB');
-      // pass "this" explicitly to enable this.timeout() calls
-      await helperDb.before.call(this);
+      await helperDb.before();
 
       debug('before(): create tmp dir for config.filesRoot');
       const tmpDir = await tmp.dir({ unsafeCleanup: true });
@@ -67,7 +66,9 @@ export function before(courseDir: string | string[] = TEST_COURSE_PATH): () => P
       load.initEstimator('python', 1);
 
       debug('before(): initialize code callers');
-      await codeCaller.init();
+      await codeCaller.init({ lazyWorkers: true });
+
+      debug('before(): initialize assets');
       await assets.init();
 
       debug('before(): start server');
@@ -126,7 +127,7 @@ export async function after(): Promise<void> {
     await cache.close();
 
     debug('after(): finish DB');
-    await helperDb.after.call(this);
+    await helperDb.after();
   } finally {
     debug('after(): complete');
   }
@@ -134,7 +135,6 @@ export async function after(): Promise<void> {
 
 export async function waitForJobSequence(job_sequence_id) {
   let job_sequence;
-  // eslint-disable-next-line no-constant-condition
   while (true) {
     const result = await sqldb.queryOneRowAsync(sql.select_job_sequence, {
       job_sequence_id,