+++ apps/prairielearn/src/tests/honorCode.test.ts
@@ -1,31 +1,26 @@
-import { assert } from 'chai';
-
-import * as sqldb from '@prairielearn/postgres';
+import { afterAll, assert, beforeAll, describe, it } from 'vitest';
 
 import { config } from '../lib/config.js';
+import { selectAssessmentByTid } from '../models/assessment.js';
 
 import * as helperClient from './helperClient.js';
 import * as helperServer from './helperServer.js';
 
-const sql = sqldb.loadSqlEquiv(import.meta.url);
-
 describe('Exam assessment response to `requireHonorCode`', function () {
-  this.timeout(60000);
-
   const context: Record<string, any> = {};
   context.siteUrl = `http://localhost:${config.serverPort}`;
   context.baseUrl = `${context.siteUrl}/pl`;
   context.courseInstanceBaseUrl = `${context.baseUrl}/course_instance/1`;
 
-  before('set up testing server', helperServer.before());
+  beforeAll(helperServer.before());
 
-  after('shut down testing server', helperServer.after);
+  afterAll(helperServer.after);
 
   it('visits the landing page of default assessment', async () => {
-    const results = await sqldb.queryOneRowAsync(sql.select_exam, {
-      number: '1',
+    const { id: assessmentId } = await selectAssessmentByTid({
+      course_instance_id: '1',
+      tid: 'exam1-automaticTestSuite',
     });
-    const assessmentId = results.rows[0].id;
     const assessmentUrl = `${context.courseInstanceBaseUrl}/assessment/${assessmentId}/`;
 
     const response = await helperClient.fetchCheerio(assessmentUrl);
@@ -38,10 +33,10 @@ describe('Exam assessment response to `requireHonorCode`', function () {
   });
 
   it('visits landing page of assessment with disabled honor code', async () => {
-    const results = await sqldb.queryOneRowAsync(sql.select_exam, {
-      number: '13',
+    const { id: assessmentId } = await selectAssessmentByTid({
+      course_instance_id: '1',
+      tid: 'exam13-disableHonorCode',
     });
-    const assessmentId = results.rows[0].id;
     const assessmentUrl = `${context.courseInstanceBaseUrl}/assessment/${assessmentId}/`;
 
     const response = await helperClient.fetchCheerio(assessmentUrl);