+++ packages/config/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @prairielearn/config
 
+## 3.1.1
+
+### Patch Changes
+
+- be4444e: Upgrade all JavaScript dependencies
+- Updated dependencies [be4444e]
+  - @prairielearn/aws-imds@2.0.14
+
+## 3.1.0
+
+### Minor Changes
+
+- d3fb328: Add `reset()` method to `ConfigLoader`
+
+## 3.0.14
+
+### Patch Changes
+
+- cec09b5: Upgrade all JavaScript dependencies
+- Updated dependencies [cec09b5]
+  - @prairielearn/aws-imds@2.0.13
+
 ## 3.0.13
 
 ### Patch Changes