+++ packages/eslint-plugin-prairielearn/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/eslint-plugin
 
+## 2.1.5
+
+### Patch Changes
+
+- be4444e: Upgrade all JavaScript dependencies
+
+## 2.1.4
+
+### Patch Changes
+
+- cec09b5: Upgrade all JavaScript dependencies
+
 ## 2.1.3
 
 ### Patch Changes