+++ packages/node-metrics/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/node-metrics
 
+## 2.0.2
+
+### Patch Changes
+
+- be4444e: Upgrade all JavaScript dependencies
+
+## 2.0.1
+
+### Patch Changes
+
+- cec09b5: Upgrade all JavaScript dependencies
+
 ## 2.0.0
 
 ### Major Changes