+++ packages/postgres/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/postgres
 
+## 2.1.13
+
+### Patch Changes
+
+- be4444e: Upgrade all JavaScript dependencies
+
+## 2.1.12
+
+### Patch Changes
+
+- cec09b5: Upgrade all JavaScript dependencies
+
 ## 2.1.11
 
 ### Patch Changes