+++ packages/prettier-plugin-sql/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/prettier-plugin-sql
 
+## 2.1.6
+
+### Patch Changes
+
+- be4444e: Upgrade all JavaScript dependencies
+
+## 2.1.5
+
+### Patch Changes
+
+- cec09b5: Upgrade all JavaScript dependencies
+
 ## 2.1.4
 
 ### Patch Changes