+++ packages/run/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/run
 
+## 1.0.10
+
+### Patch Changes
+
+- be4444e: Upgrade all JavaScript dependencies
+
+## 1.0.9
+
+### Patch Changes
+
+- cec09b5: Upgrade all JavaScript dependencies
+
 ## 1.0.8
 
 ### Patch Changes